Extract testing of Hindler-Milner.
